Refactoring new code
