+++ libs/sdmx-toolkit/src/api/hierarchy-api.ts
@@ -3,10 +3,10 @@ import { StructuralMetaData } from '../models/structural-metadata/structural-met
 import { SdmxApiClient } from './sdmx-api-client';
 
 const buildGlossaryUrl = (agency: string, id: string, version: string) =>
-  `/structure/glossary/${agency}/${id}/${version}?references=hierarchy&detail=allcompletestubs`;
+  `structure/glossary/${agency}/${id}/${version}?references=hierarchy&detail=allcompletestubs`;
 
 const buildHierarchyUrl = (agency: string, id: string, version: string) =>
-  `/structure/hierarchy/${agency}/${id}/${version}?references=descendants&detail=full`;
+  `structure/hierarchy/${agency}/${id}/${version}?references=descendants&detail=full`;
 
 export class HierarchyApi {
   constructor(private client: SdmxApiClient) {}
@@ -18,7 +18,7 @@ export class HierarchyApi {
     return this.client.request<StructuralMetaData>(
       buildGlossaryUrl(agency ?? '', id ?? '', version ?? ''),
       { method: 'GET' },
-      this.client.config.sdmxPlusUrl || this.client.config.apiUrl,
+      this.client.config.sdmxPlusUrl,
     );
   }
 
@@ -27,7 +27,7 @@ export class HierarchyApi {
     return this.client.request<StructuralMetaData>(
       buildHierarchyUrl(agency ?? '', id ?? '', version ?? ''),
       { method: 'GET' },
-      this.client.config.sdmxPlusUrl || this.client.config.apiUrl,
+      this.client.config.sdmxPlusUrl,
     );
   }
 }
